+++ ICD.Connect.Devices.CrestronSPlus/Properties/AssemblyInfo.cs
@@ -4,5 +4,4 @@
 [assembly: AssemblyCompany("ICD Systems")]
 [assembly: AssemblyProduct("ICD.Connect.Devices.CrestronSPlus")]
 [assembly: AssemblyCopyright("Copyright © ICD Systems 2021")]
-[assembly: AssemblyVersion("14.0.1.0")]
-
+[assembly: AssemblyVersion("15.0.0.0")]